4step/requirements.txt
# Core geospatial libraries
geopandas>=0.13.0
osmnx>=1.6.0
shapely>=2.0.0
pandas>=2.0.0
numpy>=1.24.0

# Visualization
folium>=0.14.0
matplotlib>=3.7.0

# Configuration
PyYAML>=6.0

# Optional for performance
rtree>=1.0.0  # Spatial indexing (optional but recommended)
pyproj>=3.5.0  # Coordinate transformations